+++ src/app/core/storage.ts
@@ -5,7 +5,7 @@ import {
   Uuid,
   UuidMap
 } from 'src/types';
-import { log } from '../../utils/console';
+import { log, logError } from '../../utils/console';
 import { Release } from '../release';
 import { Music } from '../music';
 import { hasOwnProperty, isObject } from '../../utils/utils';
@@ -163,13 +163,14 @@ export class Storage {
 
       const obj = storageDataMap[key];
 
+      log('[Storage]', obj);
       try {
         uuidMap[key] = Release.fromStorageObject(obj);
       } catch (error) {
         try {
           uuidMap[key] = Music.fromObject(obj);
         } catch (error) {
-          log('[Storage] Broken storage object.', JSON.stringify(error), obj);
+          logError('[Storage] Broken storage object.', error);
         }
       }
     }