+++ src/utils/utils.ts
@@ -113,6 +113,13 @@ export function replaceTokens(
   return template;
 }
 
+/**
+ * Splits a string into an array of substrings using specified delimiters, trims whitespace, and removes empty entries.
+ *
+ * @param inputString - The string to be split.
+ * @param delimiters - A regular expression or string specifying the delimiter(s) to use for splitting.
+ * @returns An array of non-empty, trimmed substrings resulting from the split operation.
+ */
 export function splitString(
   inputString: string,
   delimiters: RegExp | string